+++ doc/Extensions/Component.md
@@ -70,24 +70,10 @@ Because of this all fields of the component table are reserved, they cannot be u
 
 # <a name="using">Using Components</a>
 
-To use components in you application, first you need to include the code.
-
-From a Discovery/Poller module:
+Create an instance of the component class:
 
 ```php
-require_once 'includes/component.php';
-```
-
-From the html tree:
-
-```php
-require_once "../includes/component.php";
-```
-
-Once the code is loaded, create an instance of the component class:
-
-```php
-$COMPONENT = new component();
+$COMPONENT = new LibreNMS\Component();
 ```
 
 ## <a name="get">Retrieving Components</a>
@@ -144,7 +130,7 @@ Options can be supplied to `getComponents` to influence which and how components
 You can filter on any of the [reserved](#reserved) fields. Filters are created in the following format:
 
 ```php
-$OPTIONS['filter'][FIELD] = array ('OPERATOR', 'CRITERIA');
+$options['filter']['FIELD'] = array ('OPERATOR', 'CRITERIA');
 ```
 
 Where:
