+++ html/includes/authenticate.inc.php
@@ -2,7 +2,6 @@
 
 @ini_set('session.use_only_cookies', 1);
 @ini_set('session.cookie_httponly', 1);
-require 'lib/PasswordHash.php';
 
 session_start();
