+++ html/includes/graphs/device/ntp_dispersion.inc.php
@@ -11,8 +11,7 @@
  * the source code distribution for details.
  */
 
-require_once "../includes/component.php";
-$component = new component();
+$component = new LibreNMS\Component();
 $options = array();
 $options['filter']['type'] = array('=','ntp');
 $components = $component->getComponents($device['device_id'], $options);
