+++ html/includes/object-cache.inc.php
@@ -1,10 +1,10 @@
 <?php
 
-require_once $config['install_dir'].'/includes/object-cache.inc.php';
-
 // FIXME queries such as the one below should probably go into index.php?
 // FIXME: This appears to keep a complete cache of device details in memory for every page load.
 // It would be interesting to know where this is used.  It probably should have its own API.
+use LibreNMS\ObjectCache;
+
 foreach (dbFetchRows('SELECT * FROM `devices` ORDER BY `hostname`') as $device) {
     $cache['devices']['hostname'][$device['hostname']] = $device['device_id'];
     $cache['devices']['id'][$device['device_id']]      = $device;
@@ -12,9 +12,9 @@ foreach (dbFetchRows('SELECT * FROM `devices` ORDER BY `hostname`') as $device) 
     $cache['device_types'][$device['type']]++;
 }
 
-$devices  = new ObjCache('devices');
-$ports    = new ObjCache('ports');
-$services = new ObjCache('services');
+$devices  = new ObjectCache('devices');
+$ports    = new ObjectCache('ports');
+$services = new ObjectCache('services');
 
 if ($devices['down']) {
     $devices['bgcolour'] = '#ffcccc';
