+++ html/index.php
@@ -66,13 +66,10 @@ require_once '../includes/definitions.inc.php';
 require '../includes/functions.php';
 require 'includes/functions.inc.php';
 require 'includes/vars.inc.php';
-require 'includes/plugins.inc.php';
-
-use LibreNMS\Plugins;
 
 $config['memcached']['ttl'] = $config['time']['now']+300;
 
-Plugins::start();
+LibreNMS\Plugins::start();
 
 $runtime_start = microtime(true);
