+++ html/pages/device/overview.inc.php
@@ -32,9 +32,8 @@ echo('
     <div class="col-md-6">
 ');
 require 'includes/dev-overview-data.inc.php';
-use LibreNMS\Plugins;
 
-Plugins::call('device_overview_container', array($device));
+LibreNMS\Plugins::call('device_overview_container', array($device));
 
 require 'overview/ports.inc.php';
 echo('
