+++ html/pages/device/routing/cisco-otv.inc.php
@@ -1,7 +1,6 @@
 <?php
 
-require_once "../includes/component.php";
-$component = new component();
+$component = new LibreNMS\Component();
 $options = array();
 $options['filter']['ignore'] = array('=',0);
 $options['type'] = 'Cisco-OTV';
