+++ html/pages/routing/cisco-otv.inc.php
@@ -1,7 +1,6 @@
 <?php
 
-require_once "../includes/component.php";
-$COMPONENT = new component();
+$COMPONENT = new LibreNMS\Component();
 $options = array();
 $options['filter']['ignore'] = array('=',0);
 $options['type'] = 'Cisco-OTV';
