+++ includes/discovery/cisco-cbqos.inc.php
@@ -15,8 +15,7 @@ if ($device['os_group'] == 'cisco') {
 
     $module = 'Cisco-CBQOS';
 
-    require_once 'includes/component.php';
-    $component = new component();
+    $component = new LibreNMS\Component();
     $components = $component->getComponents($device['device_id'],array('type'=>$module));
 
     // We only care about our device id.
