+++ includes/discovery/cisco-otv.inc.php
@@ -59,8 +59,7 @@ if ($device['os_group'] == 'cisco') {
 
     $module = 'Cisco-OTV';
 
-    require_once 'includes/component.php';
-    $component = new component();
+    $component = new LibreNMS\Component();
     $components = $component->getComponents($device['device_id'],array('type'=>$module));
 
     // We only care about our device id.
