+++ includes/discovery/ntp/cisco.inc.php
@@ -13,8 +13,7 @@
 
 $module = 'ntp';
 
-require_once 'includes/component.php';
-$component = new component();
+$component = new LibreNMS\Component();
 $components = $component->getComponents($device['device_id'],array('type'=>$module));
 
 // We only care about our device id.
