+++ includes/polling/cisco-otv.inc.php
@@ -59,8 +59,7 @@ if ($device['os_group'] == "cisco") {
 
     $module = 'Cisco-OTV';
 
-    require_once 'includes/component.php';
-    $component = new component();
+    $component = new LibreNMS\Component();
     $options['filter']['type'] = array('=',$module);
     $options['filter']['disabled'] = array('=',0);
     $components = $component->getComponents($device['device_id'],$options);
